Bug fixing.

+++ include/states/password_manager_state.hpp
@@ -112,8 +112,8 @@ namespace kmanager::state{
             bool repaint_passwords;
             
             // Constants
-            const int16_t label_width = this -> host -> host -> width() * 0.2482f;
-            const int16_t label_height{ 50.f };
+            const float label_width = this -> host -> host -> width() * 0.2482f;
+            const float label_height{ 50.f };
             const QString label_settings{ "background-color : #4682b4; color: white; padding: 8%; font-size: 20px" };
             const QString label_list_settings{ "padding: 8%" };
             const std::string username{ std::getenv( "USERNAME" ) };
